also include AdjustCssLengthProperty
